[PUBDEV-6510] further refactor xgboost utils
