+++ h2o-extensions/xgboost/src/main/java/hex/tree/xgboost/XGBoostUtils.java
@@ -1,22 +1,21 @@
 package hex.tree.xgboost;
 
 import hex.DataInfo;
+import hex.tree.xgboost.matrix.DenseMatrixFactory;
 import hex.tree.xgboost.matrix.SparseMatrixFactory;
+import hex.tree.xgboost.util.FeatureScore;
 import ml.dmlc.xgboost4j.java.DMatrix;
 import ml.dmlc.xgboost4j.java.XGBoostError;
-import ml.dmlc.xgboost4j.java.util.BigDenseMatrix;
-import water.H2O;
-import water.LocalMR;
-import water.MrFun;
 import water.fvec.Chunk;
 import water.fvec.Frame;
 import water.fvec.Vec;
 import water.util.Log;
 import water.util.VecUtils;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 
-import static hex.tree.xgboost.matrix.MatrixFactoryUtils.setResponseAndWeight;
 import static water.H2O.technote;
 import static water.MemoryManager.malloc4f;
 
@@ -80,22 +79,10 @@ public class XGBoostUtils {
         }
         if (sparse) {
             Log.debug("Treating matrix as sparse.");
-            // truly sparse matrix - no categoricals
-            // collect all nonzeros column by column (in parallel), then stitch together into final data structures
             trainMat = SparseMatrixFactory.csr(f, chunks, weightVec, responseVec, di, resp, weights);
         } else {
             Log.debug("Treating matrix as dense.");
-            BigDenseMatrix data = null;
-            try {
-                data = allocateDenseMatrix(nRows, di);
-                long actualRows = denseChunk(data, chunks, nRowsByChunk, f, weightVec, responseVec, di, resp, weights);
-                assert data.nrow == actualRows;
-                trainMat = new DMatrix(data, Float.NaN);
-            } finally {
-                if (data != null) {
-                    data.dispose();
-                }
-            }
+            trainMat = DenseMatrixFactory.dense(f, chunks, nRows, nRowsByChunk, weightVec, responseVec, di, resp, weights);
         }
 
         assert trainMat.rowNum() == nRows;
@@ -157,7 +144,7 @@ public class XGBoostUtils {
                 Log.debug("Treating matrix as sparse.");
                 trainMat = SparseMatrixFactory.csr(chunks, -1, response, di, resp, null);
             } else {
-                trainMat = dense(chunks, di, response, resp, null);
+                trainMat = DenseMatrixFactory.dense(chunks, di, response, resp, null);
             }
         } catch (NegativeArraySizeException e) {
             throw new IllegalArgumentException(technote(11,
@@ -168,151 +155,6 @@ public class XGBoostUtils {
         resp = Arrays.copyOf(resp, len);
         trainMat.setLabel(resp);
         return trainMat;
-    }
-
-    /****************************************************************************************************************
-     ************************************** DMatrix creation for dense matrices *************************************
-     ****************************************************************************************************************/
-
-    private static DMatrix dense(Chunk[] chunks, DataInfo di, int respIdx, float[] resp, float[] weights) throws XGBoostError {
-        Log.debug("Treating matrix as dense.");
-        BigDenseMatrix data = null;
-        try {
-            data = allocateDenseMatrix(chunks[0].len(), di);
-            long actualRows = denseChunk(data, chunks, respIdx, di, resp, weights);
-            assert actualRows == data.nrow;
-            return new DMatrix(data, Float.NaN);
-        } finally {
-            if (data != null) {
-                data.dispose();
-            }
-        }
-    }
-    
-    private static long denseChunk(BigDenseMatrix data,
-                                   int[] chunks, int[] nRowsByChunk, Frame f, Vec weightsVec, Vec respVec, DataInfo di,
-                                   float[] resp, float[] weights) {
-        int[] offsets = new int[nRowsByChunk.length + 1];
-        for (int i = 0; i < chunks.length; i++) {
-            offsets[i + 1] = nRowsByChunk[i] + offsets[i];
-        }
-        WriteDenseChunkFun writeFun = new WriteDenseChunkFun(f, chunks, offsets, weightsVec, respVec, di, data, resp, weights);
-        H2O.submitTask(new LocalMR(writeFun, chunks.length)).join();
-        return writeFun.getTotalRows();
-    }
-
-    private static class WriteDenseChunkFun extends MrFun<WriteDenseChunkFun> {
-        private final Frame _f;
-        private final int[] _chunks;
-        private final int[] _offsets;
-        private final Vec _weightsVec;
-        private final Vec _respVec;
-        private final DataInfo _di;
-        private final BigDenseMatrix _data;
-        private final float[] _resp;
-        private final float[] _weights;
-
-        // OUT
-        private int[] _nRowsByChunk;
-
-        private WriteDenseChunkFun(Frame f, int[] chunks, int[] offsets, Vec weightsVec, Vec respVec, DataInfo di,
-                                   BigDenseMatrix data, float[] resp, float[] weights) {
-            _f = f;
-            _chunks = chunks;
-            _offsets = offsets;
-            _weightsVec = weightsVec;
-            _respVec = respVec;
-            _di = di;
-            _data = data;
-            _resp = resp;
-            _weights = weights;
-            _nRowsByChunk = new int[chunks.length];
-        }
-
-        @Override
-        protected void map(int id) {
-            final int chunkIdx = _chunks[id];
-            Chunk[] chks = new Chunk[_f.numCols()];
-            for (int c = 0; c < chks.length; c++) {
-                chks[c] = _f.vec(c).chunkForChunkIdx(chunkIdx);
-            }
-            Chunk weightsChk = _weightsVec != null ? _weightsVec.chunkForChunkIdx(chunkIdx) : null;
-            Chunk respChk = _respVec.chunkForChunkIdx(chunkIdx);
-            long idx = _offsets[id] * _data.ncol;
-            int actualRows = 0;
-            for (int i = 0; i < chks[0]._len; i++) {
-                if (weightsChk != null && weightsChk.atd(i) == 0) continue;
-
-                idx = writeDenseRow(_di, chks, i, _data, idx);
-                _resp[_offsets[id] + actualRows] = (float) respChk.atd(i);
-                if (weightsChk != null) {
-                    _weights[_offsets[id] + actualRows] = (float) weightsChk.atd(i);
-                }
-
-                actualRows++;
-            }
-            assert idx == (long) _offsets[id + 1] * _data.ncol;
-            _nRowsByChunk[id] = actualRows;
-        }
-
-        private long getTotalRows() {
-            long totalRows = 0;
-            for (int r : _nRowsByChunk) {
-                totalRows += r;
-            }
-            return totalRows;
-        }
-
-    }
-
-    private static long denseChunk(BigDenseMatrix data, Chunk[] chunks, int respIdx, DataInfo di, float[] resp, float[] weights) {
-        long idx = 0;
-        long actualRows = 0;
-        int rwRow = 0;
-        for (int i = 0; i < chunks[0]._len; i++) {
-
-            idx = writeDenseRow(di, chunks, i, data, idx);
-            actualRows++;
-
-            rwRow = setResponseAndWeight(chunks, respIdx, -1, resp, weights, rwRow, i);
-        }
-        assert (long) data.nrow * data.ncol == idx;
-        return actualRows;
-    }
-
-    private static long writeDenseRow(DataInfo di, Chunk[] chunks, int rowInChunk,
-                                      BigDenseMatrix data, long idx) {
-        for (int j = 0; j < di._cats; ++j) {
-            int len = di._catOffsets[j+1] - di._catOffsets[j];
-            double val = chunks[j].isNA(rowInChunk) ? Double.NaN : chunks[j].at8(rowInChunk);
-            int pos = di.getCategoricalId(j, val) - di._catOffsets[j];
-            for (int cat = 0; cat < len; cat++)
-                data.set(idx + cat, 0f); // native memory => explicit zero-ing is necessary
-            data.set(idx + pos, 1f);
-            idx += len;
-        }
-        for (int j = 0; j < di._nums; ++j) {
-            float val = chunks[di._cats + j].isNA(rowInChunk) ? Float.NaN : (float) chunks[di._cats + j].atd(rowInChunk);
-            data.set(idx++, val);
-        }
-            return idx;
-    }
-
-    /****************************************************************************************************************
-     *********************************** DMatrix creation for sparse (CSR) matrices *********************************
-     ****************************************************************************************************************/
-
-    
-    /**
-     * Allocated an exactly-sized float[] array serving as a backing array for XGBoost's {@link DMatrix}.
-     * The backing array created by this method does not contain any actual data and needs to be filled.
-     *
-     * @param rowCount Number of rows to allocate data for
-     * @param dataInfo An instance of {@link DataInfo}
-     * @return An exactly-sized Float[] backing array for XGBoost's {@link DMatrix} to be filled with data.
-     */
-    private static BigDenseMatrix allocateDenseMatrix(final int rowCount, final DataInfo dataInfo) {
-        return new BigDenseMatrix(rowCount, dataInfo.fullN());
     }
 
     public static FeatureProperties assembleFeatureNames(final DataInfo di) {
@@ -372,21 +214,6 @@ public class XGBoostUtils {
             }
         }
         return featureScore;
-    }
-
-    static class FeatureScore {
-        static final String GAIN_KEY = "gain";
-        static final String COVER_KEY = "cover";
-        
-        int _frequency;
-        float _gain;
-        float _cover;
-
-        void add(FeatureScore fs) {
-            _frequency += fs._frequency;
-            _gain += fs._gain;
-            _cover += fs._cover;
-        }
     }
 
 }
